requirements.txt
fastapi
uvicorn[standard]
pydantic>=2
pydantic-settings
sqlalchemy>=2
psycopg[binary]
asyncpg
aiosqlite
alembic
redis>=5
loguru
python-dotenv
prometheus-client
opentelemetry-sdk
opentelemetry-instrumentation-fastapi
opentelemetry-exporter-otlp
httpx
ruff
black
pytest
pytest-asyncio
faker
fakeredis
passlib[bcrypt]
jinja2
